Test_Crm/backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# ── Settings ───────────────────────────────────────────────────
pydantic-settings==2.3.0

# ── Rate Limiting ──────────────────────────────────────────────
slowapi==0.1.9

# ── Logging ────────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Utilities ──────────────────────────────────────────────────
python-multipart==0.0.9

# ── Testing ────────────────────────────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0

# ── Auth ───────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
